Reorganize AI components and add Nina sheet UI
Move AI context, store and message providers into components/ai/context;
introduce typed store state and helpers; add Nina sheet pieces
(header, main, input, entry, audio toolbar, history, etc.); update
imports across the app and add i18n strings for prompt suggestions
