+++ apps/www/app/[locale]/(app)/(shared)/(main)/(learn)/exercises/[category]/[type]/[material]/[...slug]/single.tsx
@@ -9,6 +9,8 @@ import { FOUNDER } from "@repo/seo/json-ld/constants";
 import { LearningResourceJsonLd } from "@repo/seo/json-ld/learning-resource";
 import type { Locale } from "next-intl";
 import { getTranslations } from "next-intl/server";
+import { ExerciseAttempt } from "@/app/[locale]/(app)/(shared)/(main)/(learn)/exercises/[category]/[type]/[material]/[...slug]/attempt";
+import type { ExerciseRouteData } from "@/app/[locale]/(app)/(shared)/(main)/(learn)/exercises/[category]/[type]/[material]/[...slug]/data";
 import { DeferredAiSheetOpen } from "@/components/ai/deferred-sheet-open";
 import { DeferredComments } from "@/components/comments/deferred";
 import { ExerciseEntry } from "@/components/exercise/entry";
@@ -22,8 +24,6 @@ import {
   LayoutMaterialToc,
 } from "@/components/shared/layout-material";
 import { getOgUrl } from "@/lib/utils/metadata";
-import { ExerciseAttempt } from "./attempt";
-import type { ExerciseRouteData } from "./data";
 
 /** Renders the standalone single-exercise variant for one learn route. */
 export async function SingleExercisePage({
@@ -109,7 +109,9 @@ export async function SingleExercisePage({
           <LayoutMaterialFooter>
             <DeferredComments slug={data.exerciseFilePath} />
           </LayoutMaterialFooter>
-          <DeferredAiSheetOpen />
+          <DeferredAiSheetOpen
+            contextTitle={data.exercise.question.metadata.title}
+          />
         </LayoutMaterialContent>
         <LayoutMaterialToc
           chapters={{
