+++ apps/www/components/ai/chat-pagination-trigger.tsx
@@ -3,7 +3,7 @@
 import { CHAT_MESSAGES_PAGE_SIZE } from "@repo/backend/convex/chats/constants";
 import { Intersection } from "@repo/design-system/components/ui/intersection";
 import { memo } from "react";
-import { useCurrentChat } from "./chat-provider";
+import { useCurrentChat } from "@/components/ai/context/use-current-chat";
 
 export const AiChatPaginationTrigger = memo(() => {
   const status = useCurrentChat((state) => state.messageStatus);
