+++ apps/www/components/ai/context/use-ai.tsx
@@ -4,26 +4,28 @@ import { type ReactNode, useState } from "react";
 import { createContext, useContextSelector } from "use-context-selector";
 import { useStore } from "zustand";
 import { useShallow } from "zustand/react/shallow";
-import { type AiStore, createAiStore } from "@/lib/store/ai";
+import { type AiStoreApi, createAiStore } from "@/components/ai/store/create";
+import type { AiStore } from "@/components/ai/store/types";
 
-type AiStoreApi = ReturnType<typeof createAiStore>;
+const AiContext = createContext<AiStoreApi | null>(null);
 
-export const AiContext = createContext<AiStoreApi | null>(null);
-
+/** Provides the Nina store to AI components. */
 export function AiContextProvider({ children }: { children: ReactNode }) {
   const [store] = useState(() => createAiStore());
 
   return <AiContext.Provider value={store}>{children}</AiContext.Provider>;
 }
 
+/** Reads the Nina store instance from context. */
 function useAiContext() {
-  const ctx = useContextSelector(AiContext, (context) => context);
-  if (!ctx) {
-    throw new Error("useAi must be used within a AiContextProvider");
+  const context = useContextSelector(AiContext, (value) => value);
+  if (!context) {
+    throw new Error("useAi must be used within AiContextProvider");
   }
-  return ctx;
+  return context;
 }
 
+/** Reads one selected slice of Nina UI state. */
 export function useAi<T>(selector: (state: AiStore) => T) {
   const store = useAiContext();
   return useStore(store, useShallow(selector));
