+++ apps/www/components/ai/message-loading.tsx
@@ -2,8 +2,8 @@
 
 import { TypingLoader } from "@repo/design-system/components/ui/typing-loader";
 import { memo } from "react";
-import { useChat } from "@/lib/context/use-chat";
-import { useMessage } from "./message-context";
+import { useChat } from "@/components/ai/context/use-chat";
+import { useMessage } from "@/components/ai/context/use-message";
 
 export const AiChatMessageLoading = memo(() => {
   const status = useChat((state) => state.chat.status);
