+++ apps/www/components/ai/message-part/suggestions.tsx
@@ -5,7 +5,7 @@ import type { DataPart } from "@repo/ai/schema/data-parts";
 import { HugeIcons } from "@repo/design-system/components/ui/huge-icons";
 import { useTranslations } from "next-intl";
 import { memo } from "react";
-import { useChat } from "@/lib/context/use-chat";
+import { useChat } from "@/components/ai/context/use-chat";
 
 interface Props {
   message: DataPart["suggestions"];
