+++ packages/internationalization/dictionaries/en.json
@@ -502,6 +502,13 @@
     "get-articles-error": "Couldn't load articles.",
     "new-chat-title": "What do you want to learn?",
     "new-chat-description": "Just ask me, don't hesitate!",
+    "suggestion-current-material": "this material",
+    "suggestion-explain": "Explain this",
+    "suggestion-explain-prompt": "Explain {title} in simple student-friendly language.",
+    "suggestion-example": "Show example problem",
+    "suggestion-example-prompt": "Give me an example problem about {title} with a step-by-step solution.",
+    "suggestion-summary": "Get key points",
+    "suggestion-summary-prompt": "Summarize {title} into the key ideas I need to remember.",
     "new-chat": "New chat",
     "chats": "Chats",
     "search-chats": "Search your chats...",
