+++ packages/internationalization/dictionaries/id.json
@@ -502,6 +502,13 @@
     "get-articles-error": "Gagal ambil artikel.",
     "new-chat-title": "Mau belajar apa hari ini?",
     "new-chat-description": "Tanya aku aja ya, jangan ragu-ragu!",
+    "suggestion-current-material": "materi ini",
+    "suggestion-explain": "Jelasin ini",
+    "suggestion-explain-prompt": "Jelaskan {title} dengan bahasa yang mudah dipahami siswa.",
+    "suggestion-example": "Buat contoh soal",
+    "suggestion-example-prompt": "Buatkan contoh soal tentang {title} beserta pembahasan langkah demi langkah.",
+    "suggestion-summary": "Ambil intinya",
+    "suggestion-summary-prompt": "Ringkas {title} menjadi ide penting yang perlu aku ingat.",
     "new-chat": "Obrolan baru",
     "chats": "Obrolan",
     "search-chats": "Cari obrolan kamu...",
